requirements.txt
apache-beam[gcp]
google-cloud-aiplatform==1.7.1
google-cloud-pipeline-components==0.2.1
kfp==1.8.10
# tfx==1.4.0
# pre-commit